--- app/src/main/java/com/gn41/appandroidkotlin/data/repositories/RideRepositoryImpl.kt
@@ -1,45 +1,0 @@
-package com.gn41.appandroidkotlin.data.repositories
-
-import com.gn41.appandroidkotlin.data.dto.createRide.ActiveRideDto
-import com.gn41.appandroidkotlin.core.connectivity.NetworkHelper
-import com.gn41.appandroidkotlin.data.dto.createRide.CreateRideRequestDto
-import com.gn41.appandroidkotlin.data.dto.createRide.RideUserDto
-import com.gn41.appandroidkotlin.data.services.rides.RideService
-import com.gn41.appandroidkotlin.localStorage.LocalStorageManager
-
-class RideRepositoryImpl(private val rideService: RideService,
-    private val networkHelper: NetworkHelper,
-    private val localStorageManager: LocalStorageManager
-) : RideRepository {
-    override suspend fun createRide(request: CreateRideRequestDto) : Result<Unit> {
-        return rideService.create(request)
-    }
-
-    override suspend fun cancelRide(id: Int) : Result<Unit> {
-        return rideService.cancelRide(id)
-    }
-
-    override suspend fun getActiveRide(): ActiveRideDto? {
-        return rideService.getActiveRide()
-    }
-
-    override suspend fun getRideUsers(id: Int, state: String): List<RideUserDto>? {
-        return rideService.getRideUsers(id, state)
-    }
-
-    override fun availableConnection() : Boolean {
-        return networkHelper.isInternetAvailable()
-    }
-
-    override suspend fun saveCache() {
-        localStorageManager.saveFormState()
-    }
-
-    override suspend fun readLocalStorage(): String {
-        return localStorageManager.readFormState()
-    }
-
-    override suspend fun clearLocalStorage() {
-        localStorageManager.clearFormState()
-    }
-}
