+++ app/src/main/java/com/gn41/appandroidkotlin/data/repositories/RidesRepository.kt
@@ -1,9 +1,20 @@
 package com.gn41.appandroidkotlin.data.repositories
 
 import com.gn41.appandroidkotlin.data.dto.rides.RideDto
+import com.gn41.appandroidkotlin.data.services.rides.RidesService
 
-interface RidesRepository {
-    suspend fun getRides(token: String): List<RideDto>?
-    suspend fun getRiderDriverRecommendation(riderId: Int, driverId: Int, token: String): Double?
+class RidesRepository(private val ridesService: RidesService) {
+
+    suspend fun getRides(token: String): List<RideDto>? {
+        return ridesService.getRides(token)
+    }
+
+    suspend fun getRiderDriverRecommendation(
+        riderId: Int,
+        driverId: Int,
+        token: String
+    ): Double? {
+        return ridesService.getRiderDriverRecommendation(riderId, driverId, token)
+    }
 }
