--- app/src/main/java/com/gn41/appandroidkotlin/data/repositories/UserIdRepository.kt
@@ -1,7 +1,0 @@
-package com.gn41.appandroidkotlin.data.repositories
-
-import com.gn41.appandroidkotlin.data.dto.user.UserIdDto
-
-interface UserIdRepository {
-    suspend fun getUserByAuthId(authId: String): UserIdDto
-}
