+++ app/src/main/java/com/gn41/appandroidkotlin/data/repositories/ZoneRepository.kt
@@ -1,9 +1,14 @@
 package com.gn41.appandroidkotlin.data.repositories
 
 import com.gn41.appandroidkotlin.data.dto.zone.ZoneDto
+import com.gn41.appandroidkotlin.data.services.zones.ZoneService
 
-interface ZoneRepository {
-    suspend fun getZones() : List<ZoneDto>
+class ZoneRepository(private val zoneService: ZoneService) {
+    suspend fun getZones() : List<ZoneDto> {
+        return zoneService.getZones()
+    }
 
-    suspend fun getZoneByName(name: String) : ZoneDto
+    suspend fun getZoneByName(name:String) : ZoneDto {
+        return zoneService.getZoneByName(name)
+    }
 }
