+++ app/src/main/java/com/gn41/appandroidkotlin/presentation/viewmodels/CreateRideViewModel.kt
@@ -1,6 +1,5 @@
 package com.gn41.appandroidkotlin.presentation.viewmodels
 
-import android.util.ArrayMap
 import android.util.Log
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.getValue
@@ -8,7 +7,6 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.gn41.appandroidkotlin.cache.CacheManager
-import com.gn41.appandroidkotlin.core.connectivity.NetworkHelper
 import com.gn41.appandroidkotlin.data.dto.createRide.CreateRideRequestDto
 import com.gn41.appandroidkotlin.data.dto.vehicle.VehicleDto
 import com.gn41.appandroidkotlin.data.dto.zone.ZoneDto
@@ -17,7 +15,6 @@ import com.gn41.appandroidkotlin.data.repositories.RideRepository
 import com.gn41.appandroidkotlin.data.repositories.VehicleRepository
 import com.gn41.appandroidkotlin.data.repositories.ZoneRepository
 import com.gn41.appandroidkotlin.data.services.performance.Supervisor
-import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import kotlin.time.measureTimedValue
