+++ app/src/main/java/com/gn41/appandroidkotlin/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
 package com.gn41.appandroidkotlin.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
